src/requirements.txt
pandas
python-dotenv
requests
certifi
